If the application is already open, poll for it to be closed before deleting the file.

+++ plug-in/PDFViewFile/VB.NET/PDFViewFile.vb
@@ -7,6 +7,8 @@ Imports System.Diagnostics
 
 
 Module PDFViewFile
+    Const ERROR_SHARING_VIOLATION As Integer = 32
+
     Sub Main(ByVal args() As String)
         Try
             If args.Length = 1 Then 'the only parameter is the PDF file name
@@ -18,9 +20,27 @@ Module PDFViewFile
                         }
                     }
                     process.Start()
-                    process.WaitForExit()
-                    'delete file after the viewer is closed
-                    File.Delete(args(0))
+                    If process.HasExited Then 'an existing process was already started, so poll until the file can be deleted
+
+                        Do While True
+                            Try
+                                File.Delete(args(0))
+                                Exit Do
+                            Catch ex As System.IO.IOException
+                                If (ex.HResult And &HFFFF) = ERROR_SHARING_VIOLATION Then 'keep polling while the file is open
+                                    Threading.Thread.Sleep(1000)
+                                Else
+                                    Exit Do 'exit for an unknown error
+                                End If
+                            Catch ex As Exception
+                                Exit Do
+                            End Try
+                        Loop
+                    Else
+                        process.WaitForExit()
+                        'delete file after the viewer is closed
+                        File.Delete(args(0))
+                    End If
                 End If
             Else
                 MessageBox.Show("Invalid number of parameters")
